imx: fix boot device detection following soft reset (ie bmode)
To determine the boot device after power-on-reset you must look at
SRC_SBMR1 but if you want to determine this after a soft-reset
you should look at SRC_GPR9 which is written to by the boot rom.
To make this decision SRC_GPR10[28] is examined which is also
set by the boot rom on soft reset.

+++ arch/arm/include/asm/imx-common/sys_proto.h
@@ -79,7 +79,7 @@ enum imx6_bmode {
 	IMX6_BMODE_NAND,
 };
 
-static inline u8 imx6_is_bmode_from_gpr9(void)
+static inline u32 imx6_is_bmode_from_gpr9(void)
 {
 	return readl(&src_base->gpr10) & IMX6_SRC_GPR10_BMODE;
 }
